convert integer to float
convert integer to float

+++ Inc/util.h
@@ -14,7 +14,7 @@
 
 #define MASK_8BIT 0xFF
 
-void fp_to_uint_array(float fp, uint8_t* arr, int* index);
-void merge_uint_to_fp(uint8_t* uint_arr, float* fp_arr,int float_arr_size);
+void fp_to_uint_array(uint32_t fp, uint8_t* arr, int* index);
+void merge_uint_to_fp(uint8_t* uint_arr, uint32_t* fp_arr,int float_arr_size);
 
 #endif /* UTIL_H_ */
